added denormalization setup part related code
Basically you need to run setup.sh as usual first. then go to Denormalization
folder to run sub_setup.sh. Notice that orderline table is not deleted.
order is added with new attr, o_ols, which is essentially a map with ol_number
as key and the non_key part of orderline as value(custom type)